# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Coffee Management API - staff administration, typed settings and immutable audit",
        "version": "1.5.0",
        "description": "Auth/catalog/orders/payments/inventory plus admin staff management, token revocation, last-operational-admin protection, allow-listed typed operational settings, and immutable administrative audit logging. Inventory deployment gate defaults false; historical flags immutable. Real external provider adapter remains unconfigured (503); no real KHQR/callback, refund, report or split payment."
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/staff": {
            "get": {
                "operationId": "listStaff",
                "summary": "List staff accounts with filters and pagination",
                "description": "Admin only. Search by name/email, filter by role or is_active. Ordered name ASC, id ASC.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 100
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "cashier",
                                "manager",
                                "admin"
                            ]
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "default": 25
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 10000,
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of staff members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StaffCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "createStaff",
                "summary": "Provision a new staff account",
                "description": "Admin only. Normalizes email and hashes password. Emits staff.created audit event.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateStaffRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Staff account provisioned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StaffResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/staff/{user}": {
            "parameters": [
                {
                    "name": "user",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "get": {
                "operationId": "getStaff",
                "summary": "Get single staff account details",
                "description": "Admin only. Excludes sensitive secrets.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff account details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StaffResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Staff member not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "operationId": "updateStaff",
                "summary": "Update staff account attributes",
                "description": "Admin only. Partial update of name, email, role, or is_active. Enforces token revocation policies and last-operational-admin protection.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStaffRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated staff member",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StaffResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Staff member not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Cannot demote or deactivate the last operational administrator",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Sanctum token issued by login; staff ability required. Browser session/cookie auth is not implemented."
            }
        },
        "schemas": {
...
            "Error": {
                "type": "object",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "ValidationError": {
                "type": "object",
                "required": [
                    "message",
                    "errors"
                ],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
            "StaffUser": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                    "id",
                    "name",
                    "email",
                    "role",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "cashier",
                            "manager",
                            "admin"
                        ]
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "StaffResponse": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/StaffUser"
                    }
                }
            },
            "StaffCollection": {
                "type": "object",
                "required": [
                    "data",
                    "links",
                    "meta"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StaffUser"
                        }
                    },
                    "links": {
                        "type": "object",
                        "required": [
                            "first",
                            "last",
                            "prev",
                            "next"
                        ],
                        "properties": {
                            "first": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "last": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "prev": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            },
                            "next": {
                                "type": [
                                    "string",
                                    "null"
                                ]
                            }
                        }
                    },
                    "meta": {
                        "type": "object",
                        "required": [
                            "current_page",
                            "from",
                            "last_page",
                            "path",
                            "per_page",
                            "to",
                            "total"
                        ],
                        "properties": {
                            "current_page": {
                                "type": "integer"
                            },
                            "from": {
                                "type": [
                                    "integer",
                                    "null"
                                ]
                            },
                            "last_page": {
                                "type": "integer"
                            },
                            "path": {
                                "type": "string"
                            },
                            "per_page": {
                                "type": "integer"
                            },
                            "to": {
                                "type": [
                                    "integer",
                                    "null"
                                ]
                            },
                            "total": {
                                "type": "integer"
                            },
                            "links": {
                                "type": "array",
                                "items": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "CreateStaffRequest": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                    "name",
                    "email",
                    "role",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "cashier",
                            "manager",
                            "admin"
                        ]
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 12,
                        "maxLength": 1024
                    }
                }
            },
            "UpdateStaffRequest": {
                "type": "object",
                "additionalProperties": false,
                "minProperties": 1,
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "cashier",
                            "manager",
                            "admin"
                        ]
                    },
                    "is_active": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}